adds about page
