+++ src/pages/AboutMuseum/index.jsx
@@ -1,53 +1,22 @@
-import { Box, Typography } from '@mui/material';
 import React from 'react';
+import { Box, Typography } from '@mui/material';
 
 const AboutMuseum = () => {
   return (
-    <Box>
-      <Box style={{ borderBottom: '2px solid gray', marginBottom: '35px', padding: '30px 0' }}>
-        <Typography variant="h4" gutterBottom>
-          Вартість послуг
-        </Typography>
-        <Typography variant="h5" gutterBottom>
-          Відвідування експозиції музею (І -IV поверхів)
-        </Typography>
-        <Box>
-          <Typography>- діти – 15,00 грн.</Typography>
-          <Typography>- студенти – 20,00 грн.</Typography>
-          <Typography>- пенсіонери – 20,00 грн.</Typography>
-          <Typography>- дорослі – 30,00 грн.</Typography>
-        </Box>
-      </Box>
-      <Box style={{ borderBottom: '2px solid gray', marginBottom: '35px', padding: '30px 0' }}>
-        <Typography variant="h4" gutterBottom>
-          Екскурсійне обслуговування груп від 10 осіб
-        </Typography>
-        <Typography variant="h5" gutterBottom>
-          Тематична екскурсія (45 хвилин)
-        </Typography>
-        <Box>
-          <Typography>- діти – 25,00 грн. (з вхідною платою 15,00 грн.) – 20,00 грн.</Typography>
-          <Typography>- студенти – 30 грн. (з вхідною платою 20,00 грн.) – 30,00 грн.</Typography>
-          <Typography>- пенсіонери – 30 грн. (з вхідною платою 20,00 грн.) – 30,00 грн.</Typography>
-          <Typography>- дорослі – 45 грн. (з вхідною платою 30,00 грн.) – 45,00 грн.</Typography>
-        </Box>
-      </Box>
-      <Box style={{ marginBottom: '55px', padding: '30px 0' }}>
-        <Typography variant="h4" gutterBottom>
-          Індивідуальне екскурсійне обслуговування групи від 1 особи
-        </Typography>
-        <Box>
-          <Typography>
-            - тематична індивідуальна екскурсія незалежно від вікової категорії – 60,00 грн.
-          </Typography>
-          <Typography>
-            - оглядова індивідуальна екскурсія незалежно від вікової категорії - 110,00 грн.
-          </Typography>
-          <Typography>
-            - індивідуальна екскурсія за межами музею незалежно від вікової категорії – 80,00 грн.
-          </Typography>
-        </Box>
-      </Box>
+    <Box sx={{ marginBottom: '50px' }}>
+      <Typography variant='h4'>Історія музею</Typography>
+      <p>
+        <b>Черкаський обласний краєзнавчий музей</b> – один з найстаріших музеїв України. Його створено у травні 1918 року за ініціативи місцевого осередку «Просвіти» та вчительської громадськості міста, як історико-педагогічний музей імені Тараса Григоровича Шевченка. Очолив його невтомний дослідник краю Дмитро Панасович Бочков, який перебував на посаді директора музею з 1918 по 1935 рік. Основу зібрання склали предмети побуту, знаряддя праці, документи та стародруки, які передали музею мешканці міста, а також матеріальні цінності, що були реквізовані в панських маєтках Бобринських, Балашових, Браницьких, Лопухіних, Кантакузен-Сперанських та військові реліквії Орловського і Брянського полків, які зберігались в полковій церкві. Фонди музею мали унікальні колекції творів образотворчого мистецтва, монет, орденів та медалей, зброї, військових прапорів, коштовних прикрас, які у 20-х роках ХХ сторіччя налічували понад 15 тис. предметів, а в бібліотеці було майже 13 тис. томів книг. З 1920 року музей було взято на державне утримання і підпорядковано відділу освіти Черкаського повітового виконавчого комітету Ради робітничих, селянських та червоноармійських депутатів. У цьому ж році остаточно вирішилось питання з поверненням музею у приміщення на вулиці Гоголівській, буд. 115, де він і працював впродовж багатьох наступних років.
+      </p>
+      <p>
+        Протягом 20-х років ХХ сторіччя значну кількість найбільш цінних музейних предметів було втрачено, але їх було достатньо для створення окремого фондового відділу, який розмістився у сусідньому з основним приміщенням будинку-флігелі. Там же знаходилась і наукова бібліотека музею. У 1924 р. музей мав відділи: природничий,виробничий, історико-культурний, медико-санітарний та архівно-бібліотечний; його фонди у 1926 р. налічували 16760 предметів. При музеї працював краєзнавчий гурток друзів музею. Для відвідувачів музей був відкритий з 10-ї до 15-ї години, квиток коштував 10 копійок. Екскурсії проводилися за домовленістю, за необхідності пояснення давались французькою та німецькою мовами. З введенням обласного поділу 1932 р. більшість правобережних районів сучасної Черкащини увійшли до Київської області і з цього часу музей набуває статусу міжрайонного з назвою «Черкаський міжрайонний краєзнавчий музей Придніпровських районів Правобережжя». Директором музею з 1935 по 1941 рік працює Бирюков Петро Семенович (загинув на фронті). Великих збитків завдала музею тимчасова окупація міста 1941-1943 років, коли частина фондів, що евакуювались, потрапила під бомбардування, а ще частина була захоплена німцями. Всього за роки війни було втрачено майже 4 тис. музейних предметів, серед яких були і раритети. Загальну суму втрат музею згідно державного акта від 12 квітня 1944 р. становили 698134 крб. в тодішніх грошах. Частково були зруйноване й допоміжне приміщення музею. Завдяки невтомній праці колективу, який складався всього з 10 працівників,під керівництвом Семашка Всеволода Васильовича (директора музею в 1944-1945 рр.), 1 травня 1944 року музей першим в Україні прийняв відвідувачів звільненого від фашистів краю. У 1945 році директором музею призначено демобілізованого після поранення офіцера Івана Івановича Кравця. У 1948 р. музей мав відділи природи і історії, фонди нараховували 10386 одиниць збереження, книжковий фонд становив 5727 примірників книг для службового користування. У тому ж році музей відвідало 47147 осіб, з них 13718 дорослих, 21937 - школярів, 11492 – дошкільнят. Було проведено 232 екскурсії. За цими показниками – напружені будні колективу, який у складних умовах і за короткий час зумів не тільки відновити роботу провідної музейної установи краю, а й закласти фундамент для її розширення.
+      </p>
+      <p>
+        Добрі перспективи для подальшої успішної діяльності музею відкрилися з утворенням у 1954 році Черкаської області із центром у Черкасах. Постановою Ради Міністрів УРСР від 21 липня 1954 року Черкаський міжрайонний краєзнавчий музей був реорганізований у обласний краєзнавчий музей. Однак ще тривалий час його матеріальна база і штати залишалися без змін. Лише згодом були внесені корективи до штатного розпису, виділено спецавтомашини з кіноустановкою. У 1960-ті роки музей стає науково-методичним центром краєзнавчого руху Черкащини. За допомогою музейних фахівців і методистів в районах та селах області створюються музеї та музейні кімнати. Музей в цей час очолювали – Загребельний Григорій Іванович (1960-1968 рр.), Грушкін Віктор Михайлович (1968 р.) та Антоненко Степан Семенович (1968-1970 рр.). У 1970 році колектив музею очолив Іван Данилович Ракітін, який поставив питання про будівництво нової споруди, оскільки зношеність старого приміщення музею становила більше 70 %, і виключала можливість його ремонту. З 1973 року, після закриття експозиції для відвідувачів, розпочинається новий, позначений напруженою працею, період в історії музею. В цей час музей очолює герой Радянського Союзу Митрофан Єгорович Малущенко, зусилля якого разом з заступником директора з наукової роботи Павлом Петровичем Сосою та колективом музею спрямовуються на розробку концепції майбутньої експозиції, наукове комплектування фондів, будівництво нового приміщення, написання тематико-експозиційного плану та розробку монтажних листів.
+      </p>
+      <p>
+        Напередодні визначення події – 40-ої річниці Перемоги у німецько-радянській війні – 8 травня 1985 року урочисто було відкрито нове приміщення музею. Новий обласний краєзнавчий музей розмістився в центрі міста, в одному з мальовничих його куточків, біля Пагорба Слави. Споруджений він за проектом відомих черкаських архітекторів Л. С. Кондрацького, М. Я. Собчука, С. М. Фурсенка. Науковим консультантом був О. М. Дубовий. Вся авторська група удостоєна державної премії України ім. Т. Г. Шевченка. Експозиція музею розміщена в 30-ти залах на площі 2800 км. м. Вона має розділи: природи, археології, етнографії, історії краю, зал вшанування пам*яті Тараса Григоровича Шевченка, зал голограм, виставкові зали. Експозиція музею представляє широкий діапазон тематичних розділів, де розміщені за історичною періодизацією близько 12 тисяч експонатів, широко висвітлена історія, культура та природа краю.
+      </p>
     </Box>
   );
 };
